Fixed the package declaration in the classes so that to comply with the renaming of the directory in the previous commit
