+++ groucho-lab/src/test/java/it/cnr/iasi/saks/groucho/lab/instrument/test/experiments/oscache/testIssueCache236/LRUCache_RemoveEntryUnitTest.java
@@ -15,7 +15,7 @@
  * along with GROUCHO.  If not, see <https://www.gnu.org/licenses/>
  *
  */
-package it.cnr.iasi.saks.groucho.lab.instrument.test.experiments.oscache.test;
+package it.cnr.iasi.saks.groucho.lab.instrument.test.experiments.oscache.testIssueCache236;
 
 import org.junit.Assert;
 import org.junit.Test;
